refactor: move getMirrorConfigFromClient to utils and update imports
- Moved the getMirrorConfigFromClient function from MirrorConfig.ts to utils.ts for better organization.
- Updated all relevant imports in the query files to reflect this change.
- Removed the now obsolete MirrorConfig.ts file.
