+++ nodes/Hedera/services/mirror/utils.ts
@@ -1,16 +1,26 @@
+import createClient from 'openapi-fetch';
 import { Client, LedgerId } from '@hashgraph/sdk';
+import type { paths } from '../../core/hedera-mirror';
 
-export function getMirrorNodeUrl(client?: Client): string {
+export const getMirrorConfigFromClient = (client?: Client) => {
 	if (!client?.ledgerId) throw new Error('No Ledger Id set');
 
+	let baseUrl: string;
 	switch (client.ledgerId) {
 		case LedgerId.TESTNET:
-			return 'https://testnet.mirrornode.hedera.com';
+			baseUrl = 'https://testnet.mirrornode.hedera.com';
+			break;
 		case LedgerId.PREVIEWNET:
-			return 'https://previewnet.mirrornode.hedera.com';
+			baseUrl = 'https://previewnet.mirrornode.hedera.com';
+			break;
 		case LedgerId.MAINNET:
-			return 'https://mainnet.mirrornode.hedera.com';
+			baseUrl = 'https://mainnet-public.mirrornode.hedera.com';
+			break;
 		default:
 			throw new Error('Invalid Ledger ID');
 	}
-}
+
+	const mirrorClient = createClient<paths>({ baseUrl });
+
+	return { client: mirrorClient, baseUrl };
+};
